addressed comments, added new lint ui
