+++ querybook/webapp/hooks/queryEditor/useAutoComplete.ts
@@ -1,0 +1,25 @@
+import CodeMirror from 'lib/codemirror';
+import { useMemo } from 'react';
+import {
+    AutoCompleteType,
+    SqlAutoCompleter,
+} from 'lib/sql-helper/sql-autocompleter';
+
+export function useAutoComplete(
+    metastoreId: number,
+    autoCompleteType: AutoCompleteType,
+    language: string
+) {
+    const autoCompleter = useMemo(
+        () =>
+            new SqlAutoCompleter(
+                CodeMirror,
+                language,
+                metastoreId,
+                autoCompleteType
+            ),
+        [language, metastoreId, autoCompleteType]
+    );
+
+    return autoCompleter;
+}
